@mclaughlin6464 - Fixed deprecation warning due to using float as number of bins

+++ halotools/utils/table_utils.py
@@ -50,7 +50,7 @@ def compute_prim_haloprop_bins(dlog10_prim_haloprop=0.05, **kwargs):
         num_prim_haloprop_bins = (lg10_max_prim_haloprop - lg10_min_prim_haloprop) / dlog10_prim_haloprop
         prim_haloprop_bin_boundaries = np.logspace(
             lg10_min_prim_haloprop, lg10_max_prim_haloprop,
-            num=ceil(num_prim_haloprop_bins))
+            num=int(ceil(num_prim_haloprop_bins)))
 
     # digitize the masses so that we can access them bin-wise
     # print "PHP",np.max(prim_haloprop), prim_haloprop_bin_boundaries[-1]
@@ -183,7 +183,7 @@ class compute_conditional_decorator(object):
             else:
                 same_dict = True
 
-        if same_dict: 
+        if same_dict:
             prim_haloprop_bins = self.last_prim_haloprop_bins
         else:
             prim_haloprop_bins = compute_prim_haloprop_bins(**compute_prim_haloprop_bins_dict)
